test(error-tracking): reuse bundler test helpers
